% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.structure.R
\name{cov.structure}
\alias{cov.structure}
\title{Estimate cross-covariances of two stationary multivariate time series}
\usage{
cov.structure(X, Y = NULL, q = 10)
}
\arguments{
\item{X}{a multivariate time series represented as a \eqn{T \times p_1} matrix,
where \eqn{T} is the number of observations and \eqn{p_1} is the number of covariates.}

\item{Y}{a multivariate time series represented as a \eqn{T \times p_2} matrix,
where \eqn{T} is the number of observations and \eqn{p_2} is the number of covariates.
If \eqn{Y = NULL} then \eqn{Y := X} is used and autocovariances of \eqn{X} are computed.}

\item{q}{covariances for lags \eqn{|h| \leq q}
\code{q} must be a positive integer.}
}
\value{
Function returns a time domain object (\code{\link{timedom}}) of dimensions \eqn{(2q + 1) \times p_1 \times p_2}
representing the estimated covariance structure on lags \eqn{\{ -q,...,0,...,q\}}.
}
\description{
Estimate the covariance structure of two stationary multivariate time series
on a given set of lags \eqn{[-q,q]}.
}
\details{
Given a stationary multivariate time series (\eqn{X_t}',\eqn{Y_t}')' of dimentian \eqn{p_1 + p_2}
this function determines empirical lagged covariances between the series \eqn{X_t} and \eqn{Y_t}. More
precisely, it returns
\deqn{(\hat C^{XY}(h): h \in \{ -q,...,0,...,q\})}
where \eqn{\hat C^{XY}(h)} is the empirical version of \eqn{Cov(X_{h} Y_0)}.
For a sample of size \eqn{T} we set \eqn{\mu^X = \frac{1}{T}\sum_{t=1}^T X_t} and
\eqn{\mu^Y = \frac{1}{T}\sum_{t=1}^T Y_t}. Then for \eqn{h \geq 0} we define
\deqn{\hat C^{XY}(h) = \frac{1}{T}\sum_{t=1}^{T-h} (X_{t+h} - \hat\mu^X)(Y_{t} - \hat\mu^Y)'}
and for \eqn{h < 0}
\deqn{\hat C^{XY}(h) = \frac{1}{T}\sum_{t=|h|+1}^{T} (X_{t+h} - \hat\mu^X)(Y_{t} - \hat\mu^Y)'.}

If \eqn{Y_t} is not given, we assume \eqn{Y_t := X_t} and the function will return autocovariances of \eqn{X_t}.

Use \code{\link{lagged.cov}} to estimate a particular lag.
}
\examples{
X = rar(100)
Y = rar(100)
cov.structure(X,Y)
}
\references{
Peter J. Brockwell and Richard A. Davis
\emph{Time Series: Theory and Methods}
Springer Series in Statistics, 2009
}
\seealso{
\code{\link{lagged.cov}}
}
