% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca.filters.R
\name{dpca.filters}
\alias{dpca.filters}
\title{Compute DPCA filter coefficients}
\usage{
dpca.filters(F, L = 10)
}
\arguments{
\item{X}{multivariate stationary time series}

\item{V}{correlation structure between coefficients of vectors (default diagonal)}

\item{lags}{requested filter coefficients}

\item{q}{window for spectral density estimation as in \code{\link{spectral.density}}}

\item{weights}{as in \code{\link{spectral.density}}}

\item{freq}{frequency grid to estimate on as in \code{\link{spectral.density}}}
}
\value{
principal components series
}
\description{
For a given process \code{X} eigendecompose it's spectral density
and use an inverse fourier transform to get coefficients of the optimal
filter. For details please refer to Hormann et al paper.
}
\references{
Hormann Siegfried, Kidzinski Lukasz and Hallin Marc.
\emph{Dynamic functional principal components.} Journal of the Royal
Statistical Society: Series B (Statistical Methodology) 77.2 (2015): 319-348.
}
\seealso{
\code{\link{dpca.inverse}}, \code{\link{dpca.scores}}
}
