% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca.scores.R
\name{dpca.scores}
\alias{dpca.scores}
\title{Compute scores of dynamic principal components}
\usage{
dpca.scores(X, dpcs)
}
\arguments{
\item{X}{series to 'project'}

\item{dpcs}{principal components series}
}
\value{
Matrix of scores of \code{X}
}
\description{
Compute scores of a given series \code{X} using dynamic principal components \code{dpcs}.
Procedure can be inverted using \code{\link{dpca.inverse}}
}
\references{
Hormann Siegfried, Kidzinski Lukasz and Hallin Marc.
\emph{Dynamic functional principal components.} Journal of the Royal
Statistical Society: Series B (Statistical Methodology) 77.2 (2015): 319-348.
}
\seealso{
\code{\link{dpca.inverse}}, \code{\link{dprcomp}}
}
