% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.process.R
\name{filter.process}
\alias{filter.process}
\title{Convolute (filter) a multivariate time series using a time-domain filter}
\usage{
filter.process(X, A)
}
\arguments{
\item{X}{a multivariate time series represented with a \eqn{T \times p_1} matrix,
where \eqn{T} is the number of observations and \eqn{p_1} is the number of covariates.}

\item{A}{a time-domain filter \code{\link{timedom}}, i.e. a set of linear operators \eqn{A_k \in \mathbf{R}^{p_1 \times p_2}}
defined on a set of lags \eqn{S \subset \mathbf{Z}}.}
}
\value{
A multivariate time series \eqn{Y_t} represented as a matrix \eqn{T \times p_2}.
}
\description{
Given a multivariate time series \eqn{X_t} and a time-domain filter \eqn{\{A_k : k \in S\}} defined on lags \eqn{S}, \code{filter.process} convolutes
the time series with the filter.
}
\details{
Let \eqn{X_t} be a multivariate time series and let \eqn{\{A_k : k \in S\}} be a time domain filter defined on lags \eqn{S}
For a sample of size \eqn{T} and for \eqn{t \in \{1,...,T\}}, the functions \code{filter.process} computes
\deqn{Y_t = \sum_{k \in \mathbf{Z}} A_k X_{t-k}}
assuming \eqn{X_t = 0} for \eqn{t \notin \{1,...,T\}} and \eqn{A_k = 0} for \eqn{k \notin S}.
}
\seealso{
\code{\link{timedom}}
}
