% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.inverse.R
\name{fourier.inverse}
\alias{fourier.inverse}
\title{Inverse Fourier Transform of a given frequency-domain filter.}
\usage{
fourier.inverse(F, lags = 0)
}
\arguments{
\item{F}{an object of class \code{freqdom} which is corresponding to a function with values in \eqn{\mathbf{C}^{d_1\times d_2}}. To guarantee accuracy of inversion it is important that \eqn{F\$freq} is a dense grid of frequencies in \eqn{[-\pi,\pi]}.}

\item{lags}{lags of the Fourier coefficients to be computed.}
}
\value{
An object of class \code{timedom}. The list has the following components:
\itemize{
\item \code{operators} an array. The \eqn{k}-th matrix in this array corresponds to the \eqn{k}-th Fourier coefficients.
\item \code{lags} the lags of the corresponding Fourier coefficient.
}
}
\description{
Computes Fourier coefficients of some functional represented by an object of class \code{freqdom}.
}
\details{
Consider a function \eqn{F \colon [-\pi,\pi]\to\mathbf{C}^{d_1\times d_2}}. Its \eqn{k}-th Fourier coefficient is given as
\deqn{
  \frac{1}{2\pi}\int_{-\pi}^\pi F(\omega) \exp(ik\omega)d\omega.
}
We represent the function \eqn{F} by an object of class \code{freqdom} and approximate the integral via
\deqn{
\frac{1}{|F\$freq|}\sum_{\omega\in {F\$freq}} F(\omega) \exp(i k\omega),
}
for \eqn{h\in} lags.
}
\examples{
n = 100
X = rar(n)
Y = rar(n)
#estimate regressors in model $Y_t = \\sum_{i\\in Z} A_i X_{t-i}$
SYX = spectral.density(Y, X)
SXX = spectral.density(X)
R = freqdom.ratio(SYX,SXX, n)
A = fourier.inverse(R) 
}
\seealso{
\code{\link{fourier.transform}}
}
