% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.transform.R
\name{fourier.transform}
\alias{fourier.transform}
\title{Computes the Fourier transform of a given series of operators or a given multivariate time series}
\usage{
fourier.transform(A, freq = NULL)
}
\arguments{
\item{A}{a time-domain object \code{\link{timedom}}, i.e. a set of linear operators \eqn{A_k \in \mathbf{R}^{p_1 \times p_2}}
for some set \eqn{G \subset \mathbf{Z}} of lags. Alternatively, a multivariate time series.}

\item{freq}{frequencies on which the transfom should be evaluated. A vector of increasing values in \eqn{[-\pi,\pi]}.
If the parameter is \code{NULL}, the default grid of \eqn{201} equidistributed points on \eqn{[-\pi,\pi]} is taken
as the grid. Here we refer to this set as \eqn{G}.}
}
\value{
A frequency domain operator \eqn{F = \{ F_\theta : \theta \in G \}} (\code{\link{freqdom}}) defined on the given set of frequencies \eqn{G}.
Every \eqn{F_\theta} corresponds to the Fourier transform of \eqn{A} evaluated on \eqn{\theta \in G}.
}
\description{
For a given set of frequencies \eqn{G \subset [-\pi,\pi]} and a  series of operators
\eqn{A = \{ A_k : k \in S \}} (\code{\link{timedom}}), where \eqn{S} is a set of lags,
the function \code{fourier.transform} evaluates the Fourier transform of
\eqn{A} on the set \eqn{G}.
}
\details{
Given a series of frequencies \eqn{G \subset [-\pi,\pi]} and a series of operators
\deqn{(A(h): h \in \{ -q,...,0,...,q\}),}
where \eqn{A(h) \in \mathbf{R}^{p_1 \times p_2}},
the function \code{fourier.transform} evaluates the Fourier transform on each \eqn{\theta \in G}.
More precisely, for each \eqn{\theta \in G} it computes
\deqn{ F_\theta = \sum_{h=-q}^q A(h) e^{-i\theta h}.}

It returns a frequency-domain operator \eqn{F = \{ F_\theta : \theta \in G \}}.
}
\examples{
X = rar(100,d=2)
C = cov.structure(X)
F = fourier.transform(C) # a simple spectral density estimator
Cinv = invfourier(F)
}
\seealso{
\code{\link{invfourier}}
}
