% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freqdom.inverse.R
\name{freqdom.inverse}
\alias{freqdom.inverse}
\title{Compute an inverse of a given Frequency Domain Operator}
\usage{
freqdom.inverse(S, n = NULL, Kconst = NULL, K = NULL)
}
\arguments{
\item{S}{first spectral density}

\item{n}{number of observations used for estimation - the precision of inversion is calculated using this parameter}

\item{Kconst}{constant for heuristic (as in \code{\link{reg.dim.est}})}

\item{K}{how many directions should be inverted (as in \code{\link{pseudoinverse}})}
}
\value{
Frequency Domain Operator object
}
\description{
For given Frequency Domain Operator \code{S} computes ratio \code{S^\{-1\}} at each frequency.
}
\seealso{
\code{\link{freqdom.product}}, \code{\link{freqdom.ratio}}
}

