% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqdom.kronecker.R
\name{freqdom.kronecker}
\alias{freqdom.kronecker}
\title{Compute the Kronecker product of two frequency-domain operators}
\usage{
freqdom.kronecker(F, G)
}
\arguments{
\item{F}{a frequency-domain filter of type \code{\link{freqdom}}, i.e. a set of linear operators \eqn{F_k \in \mathbf{R}^{p_1 \times p_2}}
on a discreet grid defined of \eqn{[-\pi,\pi]}.}

\item{G}{a frequency-domain filter of type \code{\link{freqdom}}, i.e. a set of linear operators \eqn{G_k \in \mathbf{R}^{q_1 \times q_2}}
on a discreet grid defined of \eqn{[-\pi,\pi]}.}
}
\value{
Function returns a frequency domain object (\code{\link{freqdom}}) of dimensions \eqn{L \times (p_1q_1) \times (p_2q_2)}, where \eqn{L} is the
size of the evaluation grid. The elements corresponds to \eqn{F_\theta \otimes G_\theta} defined above.
}
\description{
For given frequency-domain operators \code{F} and \code{G} (\code{\link{freqdom}}) the function \code{freqdom.kronecker} computes their Kronecker product frequency-wise.
}
\details{
Let \eqn{F = \{ F_\theta : \theta \in S \}}, \eqn{G = \{ G_\theta : \theta \in S \}},
where \eqn{S} is a finite grid of frequencies in \eqn{[-\pi,\pi]}, \eqn{F_\theta \in \mathbf{C}^{p_1 \times p_2}}
and \eqn{G_\theta \in \mathbf{C}^{q_1 \times q_2}}.

We define \deqn{H_\theta = F_\theta \otimes G_\theta} as a Kronecker product of \eqn{F_\theta} and \eqn{G_\theta}, i.e. \eqn{H_\theta \in \mathbf{R}^{p_1q_1\times p_2q_2}}.
Function \code{freqdom.kronecker} returns \eqn{H = \{ H_\theta : \theta \in S \}}.
}
\examples{
n = 100
X = rar(n,d=3)
Y = rar(n,d=3)
SX = spectral.density(X)
SY = spectral.density(Y)
R = freqdom.kronecker(SY,SX)
}
