% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.process.R
\name{\%c\%}
\alias{\%c\%}
\title{Convolution of a process X with an operator A.}
\usage{
A \%c\% X
}
\arguments{
\item{A}{time-domain filter \code{\link{timedom}}, i.e. a set of linear operators \eqn{A_k \in \mathbf{R}^{p_1 \times p_2}}
for some set \eqn{S \subset \mathbf{Z}} of lags.}

\item{X}{multivariate time series represented as a \eqn{T \times p_1} matrix,
where \eqn{T} is the number of observations and \eqn{p_1} is the number of covariates.}
}
\value{
Multivariate time series \eqn{Y_t} represented as a matrix \eqn{T \times p_2}.
}
\description{
Given a multivariate time series \eqn{X_t} and a time-domain filter \eqn{\{A_k : k \in S\}} for some lags \eqn{S}, \code{filter.process} convolutes
the time series with the filter. This is a convenience operator equivalent to \code{\link{filter.process}}. See \code{\link{filter.process}} for details.
}
\seealso{
\code{\link{filter.process}}, \code{\link{timedom}}
}
