% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freqdom.product.R
\name{\%*\%}
\alias{\%*\%}
\title{Frequency-wise or component-wise matrix product.}
\usage{
e1 \%*\% e2
}
\arguments{
\item{e1}{first element}

\item{e2}{second element}
}
\value{
object of the same type as e1 but with new dimensions
}
\description{
Frequency-wise or time-wise matrix product. Takes two elements
\code{freqdom} or \code{timedom} and multiplies them on
each frequency or time point. If objects of other type are provided
then the standard multiplication is applied.
}

