% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freqdom.kronecker.R
\name{\%x\%}
\alias{\%x\%}
\title{Frequency-wise or component-wise Kronecker product.}
\usage{
e1 \%x\% e2
}
\arguments{
\item{e1}{first element}

\item{e2}{second element}
}
\value{
object of the same type as e1 but with new dimensions
}
\description{
Frequency-wise or time-wise Kronecker product. Takes two elements
\code{freqdom} or \code{timedom} and applies the Kronecker product on
each frequency or time point. If objects of other type are provided
then the standard function is applied.
}

