% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.freqdom.R
\name{is.freqdom}
\alias{is.freqdom}
\title{Checks if a given object is of frequency domain type}
\usage{
is.freqdom(X)
}
\arguments{
\item{X}{any object}
}
\value{
\code{TRUE} if \code{X} is of type \code{\link{freqdom}}, \code{FALSE} otherwise
}
\description{
Checks if a given object is of frequency domain type (\code{\link{freqdom}}).
}
\seealso{
\code{\link{freqdom}}, \code{\link{timedom}}, \code{\link{is.timedom}}
}
