% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.timedom.R
\name{is.timedom}
\alias{is.timedom}
\title{Checks if a given object is of time domain type}
\usage{
is.timedom(X)
}
\arguments{
\item{X}{any object}
}
\value{
\code{TRUE} if \code{X} is of type \code{\link{timedom}}, \code{FALSE} otherwise
}
\description{
Checks if a given object is of time domain type (\code{\link{timedom}}).
}
\seealso{
\code{\link{freqdom}}, \code{\link{timedom}}, \code{\link{is.freqdom}}
}
