% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSE.R
\name{MSE}
\alias{MSE}
\title{Compute a mean square error between X and Y}
\usage{
MSE(X, Y)
}
\arguments{
\item{X}{first matrix to compare}

\item{Y}{second matrix to compare}
}
\value{
Estimated mean square error
}
\description{
Compute the mean square error between \eqn{X_t} and \eqn{Y_t}, as \deqn{\frac{1}{n} \sum_{k=1}^n \|X_k-Y_k\|^2,}
where \eqn{\|\cdot\|} denotes a Euclidian norm and \eqn{n} is the number of observations.
}

