% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.freqdom.R
\name{plot.freqdom}
\alias{plot.freqdom}
\title{Plot a frequency domain operator}
\usage{
\method{plot}{freqdom}(x, xcoef = NULL, ycoef = NULL, type = "cartesian",
  ...)
}
\arguments{
\item{x}{Frequency domain object}

\item{xcoef}{vector of horizontal coefficients to plot}

\item{ycoef}{vector of vertical coefficients to plot}

\item{type}{'cartesian' or 'polar' coordinates}

\item{...}{other parameters to \code{\link{plot}}}
}
\description{
Plot chosen coefficients of frequenct domain operator. Draws all pairs \eqn{ \{ (i,j) | i \in x, j\in y }.
}

