% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoinverse.R
\name{pseudoinverse}
\alias{pseudoinverse}
\title{Invert first K eigendirections of the matrix.}
\usage{
pseudoinverse(M, K = NULL, th = NULL)
}
\arguments{
\item{M}{matrix to solve}

\item{K}{number of directions to invert}

\item{th}{fixed treshold for eigenvalues}
}
\value{
Pseudoinverse matrix
}
\description{
Invert first \code{K} eigendirections of the matrix. If \code{K} is
not specified the functions takes direction with eigencalues grater
than given treshold \code{th}.
If \code{th} is also not specified then all direction are inverted
(equivalent to \code{\link[base]{solve}})
}
\references{
Hormann Siegfried, Kidzinski Lukasz and Kokoszka Piotr.
\emph{Estimation in functional lagged regression.}
Journal of Time Series Analysis 36.4 (2015): 541-561.
}
