% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.R
\name{rbm}
\alias{rbm}
\title{Generate brownian motions}
\usage{
rbm(n, d = 100)
}
\arguments{
\item{n}{number of observations to generate}

\item{d}{dimension (sampling at d points)}
}
\value{
n x d matrix with independent n observations
}
\description{
Generate \eqn{n} independent Brownian motions
}
\examples{
bm = rbm(100)
plot(bm)
}
\seealso{
\code{\link{rbb}}
}

