% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{rev.timedom}
\alias{rev.timedom}
\title{rev Reverts order of lags in an object of class \code{\link{timedom}}}
\usage{
\method{rev}{timedom}(x)
}
\description{
\itemize{
\item For a given \code{timedom} object \code{x}, the function \code{rev} reverts the order of lags.
}
}
\keyword{internal}
