% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.R
\name{rma}
\alias{rma}
\title{Moving avarege process}
\usage{
rma(n, d = 2, Psi = NULL, noise = c("mnormal", "mt"), sigma = NULL,
  df = 4)
}
\arguments{
\item{n}{number of observations to generate}

\item{d}{number of dimensions of the process}

\item{Psi}{time domain object describing operators}

\item{noise}{the underlying X process}
}
\description{
Generates \eqn{n} observations of a \eqn{d}-dimensional moving average process with operators A, i.e.
\deqn{ Y_t = \sum_{i \in L_A} A_i X_t },
where L_A is a set of lags on which operators A_i are defined.
}
