% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedom.norms.R
\name{timedom.norms}
\alias{timedom.norms}
\title{Compute operator norms of elements of a filter}
\usage{
timedom.norms(A, type = c("O", "I", "F", "M", "2"))
}
\arguments{
\item{A}{a time-domain filter \code{\link{timedom}}, i.e. a set of linear operators \eqn{A_k \in \mathbf{R}^{p_1 \times p_2}}
for some set \eqn{S \subset \mathbf{Z}} of lags.}

\item{type}{the matrix norm to be used as in \link{base::norm}}
}
\value{
A list with elements
\itemize{
\item \code{$norms} - a vector of the length \eqn{S} with the \eqn{k}-th covariate equal \eqn{\|A_{lags(k)}\|},
\item \code{$lags} - \code{lags} corresponding to the computed norms.
}
}
\description{
Given a time-domain filter \eqn{\{A_k : k \in S\}} (\code{\link{timedom}}) for some lags \eqn{S}, \code{timedom.norms}
returns \eqn{\{\|A_k\| : k \in S\}} for a given norm \eqn{\| \cdot \|}.
}
\examples{
d = 2

A = array(0,c(d,d,2))
A[1,,] = 2 * diag(d:1)/d
A[2,,] = 1.5 * diag(d:1)/d
OP = timedom(A,c(-2,1))
timedom.norms(OP)
}
