% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedom.R
\name{timedom}
\alias{timedom}
\title{Converts a multivariate time series or a time-domain filter into a time-domain object}
\usage{
timedom(A, lags = 1:dim(A)[1])
}
\arguments{
\item{A}{The elements \eqn{A[,,k], 1\leq k\leq K}, are real valued \eqn{(d_1\times d_2)} matrices (all of same dimension). These matrices define a linear filter.}

\item{lags}{a vector of increasing integers. It corresponds to the time lags of the filter.}
}
\value{
Returns an object of class \code{timedom}. An object of class  \code{timedom} is a list containing the following components:
\itemize{
\item \code{operators} - operators & returns the array A as given in the argument.
\item \code{lags} - returns the vector of lags as given in the argument.
}
}
\description{
Creates an object of class \code{timedom}. This object corresponds to a linear filter.
}
\details{
This class is used to describe a linear filter, i.e. a sequence of matrices, each of which correspond to a certain lag. Filters can, for example, be used to transform a sequence \eqn{(X_t)} into a new sequence \eqn{(Y_t)} by defining
\deqn{
 Y_t=\sum_k A_kX_{t-k}.
}
See \code{filter.process()}.
Formally we consider a collection \eqn{[A_1,\ldots,A_K]} of complex-valued matrices \eqn{A_k}, all of which have the same dimension \eqn{d_1\times d_2}. Moreover, we consider lags \eqn{\ell_1<\ell_2<\cdots<\ell_K}. The object this function creates corresponds to the mapping \eqn{f: \mathrm{lags}\to \mathbf{R}^{d_1\times d_2}}, where \eqn{\ell_k\mapsto A_k}.
}
\examples{
d = 2

A = array(0,c(d,d,2))
A[1,,] = 2 * diag(d:1)/d
A[2,,] = 1.5 * diag(d:1)/d
OP = timedom(A,c(-2,1))
print(OP)
}
