% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedom.trunc.R
\name{timedom.trunc}
\alias{timedom.trunc}
\title{Truncates a time-domain object to given lags}
\usage{
timedom.trunc(A, lags)
}
\arguments{
\item{A}{an object of class \code{timedom}.}

\item{lags}{a vector which contains a set of lags. These lags must be a subset of the lags defined for timedom object A. Only those lags will be kept, the other lags are removed.}
}
\value{
An object of class \code{timedom}.
}
\description{
This function removes lags from a linear filter.
}
\examples{
X = rar(100)
Y = rar(100)
#estimate regressors in model $Y_t = \\sum_{i\\in Z} A_i X_{t-i}$
A = speclagreg(X,Y)
B = timedom.trunc(A, c(-1, 2, 3))
}
