% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca.R
\name{dpca}
\alias{dpca}
\title{Apply DPCA}
\usage{
dpca(X, q = 3, thresh = NULL, Ndpc = dim(X)[2], center = TRUE, ...)
}
\arguments{
\item{X}{...}

\item{Ndpc}{...}

\item{center}{...}

\item{...}{...}

\item{lags}{..}
}
\value{
A list
}
\description{
Applies DPCA to a multivariate time series
}
\details{
Apply
}
