% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dpca.scores.R
\name{dpca.scores}
\alias{dpca.scores}
\title{Compute scores of dynamic principal components}
\usage{
dpca.scores(X, XI)
}
\arguments{
\item{X}{series to 'project'}

\item{XI}{principal components series}
}
\value{
Matrix of scores of \code{X}
}
\description{
Compute scores of a given series \code{X} using dynamic principal components \code{XI}.
Procedure can be inverted using \code{\link{dpca.inverse}}
}
\references{
Siegfried Hormann, Lukasz Kidzinski and Marc Hallin
Dynamic Functional Principal Component
Research report, 2012
}
\seealso{
\code{\link{dpca.inverse}}, \code{\link{dprcomp}}
}

