% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.process.R
\name{filter.process}
\alias{filter.process}
\title{Generate a linear process from}
\usage{
filter.process(X, A)
}
\arguments{
\item{X}{process}

\item{A}{time-domain operator series}
}
\value{
Multivariate linear process
}
\description{
Compute
\deqn{Y_t = \sum_{k=-q}^p A_k X_{t-k}}
where \eqn{X_t} is a stationary multivariate time series and \eqn{(A_k)_{-q \leq k \leq p}} is a filter.
Note that this procedure can also be used for deconvolution
\deqn{\tilde X_t = \sum_{k=-q}^p X_{t+k}} A_k'
but therefore in place of \code{A} one should it's transposition and inversed time, i.e.
\code{t(rev(A))}.
}

