% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freqdom.transpose.R
\name{freqdom.transpose}
\alias{freqdom.transpose}
\title{Transpose pointwise timedom or freqdom object}
\usage{
freqdom.transpose(x)
}
\arguments{
\item{x}{freqdom or timedom object}
}
\value{
object of the same type as x
}
\description{
Transpose the object at each frequency or lag
}

