% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.process.R
\name{\%c\%}
\alias{\%c\%}
\title{Convolution of a process X with an operator A.}
\usage{
A \%c\% X
}
\arguments{
\item{A}{\code{timedom} operators}

\item{X}{\code{timedom} multivariate time series}
}
\value{
Convoluted series of the same type as X
}
\description{
Convolution of a process X with an operator A. As in \code{\link{filter.process}}
}

