% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.freqdom.R
\name{is.freqdom}
\alias{is.freqdom}
\title{Checks if an object belongs to the class freqdom}
\usage{
is.freqdom(X)
}
\arguments{
\item{X}{some object}
}
\value{
\code{TRUE} if \code{X} is of type \code{\link{freqdom}}, \code{FALSE} otherwise
}
\description{
Checks if an object belongs to the class \code{\link{freqdom}}.
}
\seealso{
\code{\link{freqdom}}, \code{\link{timedom}}, \code{\link{is.timedom}}
}
\keyword{classes}
