% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.freqdom.R
\name{is.freqdom}
\alias{is.freqdom}
\title{Check if a given object is a frequency domain matrix}
\usage{
is.freqdom(X)
}
\arguments{
\item{X}{an object}
}
\value{
boolean
}
\description{
Checks if a given object is a frequency domain matrix.
}

