% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.timedom.R
\name{is.timedom}
\alias{is.timedom}
\title{Check if a given object is a time domain object}
\usage{
is.timedom(X)
}
\arguments{
\item{X}{an object}
}
\value{
boolean
}
\description{
Checks if a given object is a time domain filter or process.
}

