% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timedom.R
\name{timedom}
\alias{timedom}
\title{Creates a time domain operator}
\usage{
timedom(X, lags = NULL)
}
\arguments{
\item{X}{array of operator or matrix of observations}

\item{lags}{lags on which the operator is defined}
}
\value{
Time domain object
}
\description{
Creates a time domain operator
}

