% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedom.trunc.R
\name{timedom.trunc}
\alias{timedom.trunc}
\title{Truncates a time-domain object to given lags}
\usage{
timedom.trunc(A, lags)
}
\arguments{
\item{A}{a time-domain filter \code{\link{timedom}}, i.e. a set of linear operators \eqn{A_k \in \mathbf{R}^{p_1 \times p_2}}
for some set \eqn{S \subset \mathbf{Z}} of lags.}

\item{lags}{lags to which \eqn{A} should be truncated.}
}
\value{
A truncated \code{\link{timedom}} operator time series, with only operators on lags \eqn{S \cap T} present.
}
\description{
Truncates a time-domain object (\code{\link{timedom}}) to specified lags (equivalent to setting other lags to zero).
}
\details{
Let \eqn{A = \{A_k : k \in S\}} be a time domain filter for some lags \eqn{S}. \code{timedom.trunc}
truncates \eqn{S} to some subset \eqn{T} such that \eqn{S \backslash T} are implicitely
set to \eqn{0}. This operation can save space and computation time. Also, it simplfifies removal of
non-significant lags.

For a given \eqn{A} and a set of lags \eqn{T}, \code{timedom.trunc} returns a time domain object
representing \eqn{\{A_k : k \in S \cap T\}}.
}
\examples{
X = rar(100)
Y = rar(100)
#estimate regressors in model $Y_t = \\sum_{i\\in Z} A_i X_{t-i}$
A = speclagreg(X,Y)
B = timedom.trunc(A, c(-1, 2, 3))
}
